added population living difference queries (#77)

+++ src/main/java/com/napier/sem/CapitalCityReportView.java
@@ -19,13 +19,13 @@ public class CapitalCityReportView extends ReportView {
     /**
      * City Population
      */
-    public int Population;
+    public long Population;
 
     public CapitalCityReportView(ResultSet results) {
         try {
             Name = results.getString("Name");
             Country = results.getString("Country");
-            Population = results.getInt("Population");
+            Population = results.getLong("Population");
         } catch (Exception e) {
             System.out.println(e.getMessage());
             System.out.println("Failed to fetch country report view");
